Replace the old facts page with the new layout
I deleted the old facts.js page and refactored the grid-facts.js
filename,query and functions so it's regarded as the new facts page
